Reorganize codes.
